requirements.txt
gradio>=3.30
requests>=2.28
fastapi>=0.100
uvicorn>=0.23
ollama>=0.1.0
python-dotenv>=1.0
